requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary
redis==5.0.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
celery==5.3.4
python-decouple==3.8
gunicorn==21.2.0 